Refactor to submit changes as array
