+++ api/src/changeRequests/ValidationResult.ts
@@ -1,0 +1,4 @@
+export type ValidationResult = {
+    validated: boolean;
+    error?: string;
+};
