+++ api/src/dto/ChangeReqDto.ts
@@ -1,10 +1,19 @@
-import { DocType, Uuid } from "../enums";
-import { IsEnum, IsNotEmpty, IsObject, IsString } from "class-validator";
+import { Type } from "class-transformer";
+import { Uuid } from "../enums";
+import { IsNotEmpty, IsNumber, IsObject, IsString, ValidateNested } from "class-validator";
+
+export class ChangeReqItemDto {
+    @IsNotEmpty()
+    @IsNumber()
+    id: number;
+
+    @IsNotEmpty()
+    @IsObject()
+    doc: any; // Object containing full submitted / updated document
+}
 
 /**
  * Document structure for client / CMS submitted changes to existing documents or new documents.
- * After successful processing, a differential "change" document (ChangeDto) is generated by the
- * API from the UpdateDto and distributed to clients.
  */
 export class ChangeReqDto {
     @IsNotEmpty()
@@ -12,9 +21,7 @@ export class ChangeReqDto {
     reqId: Uuid;
 
     @IsNotEmpty()
-    @IsEnum(DocType)
-    type: DocType;
-
-    @IsObject()
-    doc: any; // Object containing full submitted / updated document
+    @ValidateNested({ each: true })
+    @Type(() => ChangeReqItemDto)
+    changes: ChangeReqItemDto[];
 }
