+++ cms/src/db/repositories/postRepository.ts
@@ -49,17 +49,20 @@ export class PostRepository extends BaseRepository {
         await db.docs.put(post);
 
         // Save change, which will be sent to the API later
+
+        const reqId = uuidv4();
+
         db.localChanges.put({
             status: LocalChangeStatus.Unsynced,
-            reqId: uuidv4(),
-            docId: contentId,
-            doc: content,
+            reqId,
+            docId: postId,
+            doc: post,
         });
         return db.localChanges.put({
             status: LocalChangeStatus.Unsynced,
-            reqId: uuidv4(),
-            docId: postId,
-            doc: post,
+            reqId,
+            docId: contentId,
+            doc: content,
         });
     }
