+++ cms/src/stores/localChanges.ts
@@ -1,20 +1,24 @@
 import { defineStore, storeToRefs } from "pinia";
 import {
+    AckStatus,
+    DocType,
     LocalChangeStatus,
     type LocalChange,
     type ChangeReqDto,
-    DocType,
     type ChangeReqAckDto,
-    AckStatus,
+    type Uuid,
+    type ChangeReqItemDto,
 } from "@/types";
 import { liveQuery } from "dexie";
 import { useObservable } from "@vueuse/rxjs";
-import { watch, type Ref } from "vue";
+import { watch, type Ref, computed } from "vue";
 import { LocalChangesRepository } from "@/db/repositories/localChangesRepository";
 import type { Observable } from "rxjs";
 import { useSocketConnectionStore } from "./socketConnection";
 import { socket } from "@/socket";
 import { db } from "@/db/baseDatabase";
+
+type LocalChangeCollection = { [key: Uuid]: LocalChange[] };
 
 export const useLocalChangeStore = defineStore("localChanges", () => {
     const localChangesRepository = new LocalChangesRepository();
@@ -25,31 +29,53 @@ export const useLocalChangeStore = defineStore("localChanges", () => {
         >,
     );
 
+    const unsyncedChangesByReqId = computed(() => {
+        if (!unsyncedChanges) {
+            return {};
+        }
+
+        return unsyncedChanges.value!.reduce((returnObject: LocalChangeCollection, change) => {
+            returnObject[change.reqId] = returnObject[change.reqId] || [];
+
+            returnObject[change.reqId].push(change);
+
+            return returnObject;
+        }, {});
+    });
+
     const watchForSyncableChanges = () => {
         const { isConnected } = storeToRefs(useSocketConnectionStore());
 
         watch([isConnected, unsyncedChanges], async ([isConnected, localChanges]) => {
             if (isConnected && localChanges && localChanges.length > 0) {
-                await syncLocalChangesToApi(localChanges);
+                await syncLocalChangesToApi(unsyncedChangesByReqId.value);
             }
         });
     };
 
-    const syncLocalChangesToApi = async (localChanges?: LocalChange[]) => {
-        await localChanges?.forEach(async (change) => {
-            if (change.status == LocalChangeStatus.Unsynced) {
+    const syncLocalChangesToApi = async (localChanges: LocalChangeCollection) => {
+        await Object.keys(localChanges).forEach(async (reqId) => {
+            const changesToSubmit = localChanges[reqId];
+
+            changesToSubmit.forEach(async (change) => {
                 await localChangesRepository.update(change, {
                     status: LocalChangeStatus.Syncing,
                 });
+            });
 
-                const changeReq: ChangeReqDto = {
-                    reqId: change.reqId,
-                    type: DocType.ChangeReq,
-                    doc: change.doc,
-                };
+            const changeReq: ChangeReqDto = {
+                reqId,
+                type: DocType.ChangeReq,
+                changes: localChanges[reqId].map(
+                    (change) =>
+                        ({
+                            id: change.id,
+                            doc: change.doc,
+                        }) as ChangeReqItemDto,
+                ),
+            };
 
-                socket.emit("data", changeReq);
-            }
+            socket.emit("data", changeReq);
         });
     };
